third_party/blink/web_tests/external/wpt/html/canvas/element/manual/wide-gamut-canvas/canvas-display-p3-pattern-image-expected.txt

This is a testharness.js-based test.
Found 8 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Display-P3-FF0000CC.png, Context srgb, ImageData display-p3
  assert_true: Actual pixel value 255,57,40,204 is approximately equal to 234,51,35,204. expected true got false
[FAIL] Adobe-RGB-FF0000CC.png, Context srgb, ImageData display-p3
  assert_true: Actual pixel value 255,61,43,204 is approximately equal to 234,51,35,204. expected true got false
[FAIL] Adobe-RGB-FF0000CC.png, Context display-p3, ImageData srgb
  assert_true: Actual pixel value 255,2,0,204 is approximately equal to 255,19,11,204. expected true got false
[FAIL] Generic-CMYK-FF000000.jpg, Context display-p3, ImageData display-p3
  assert_true: Actual pixel value 73,162,214,255 is approximately equal to 0,160,213,255. expected true got false
[FAIL] Generic-CMYK-BE000000.jpg, Context display-p3, ImageData display-p3
  assert_true: Actual pixel value 80,177,219,255 is approximately equal to 65,177,219,255. expected true got false
[FAIL] Display-P3-FFFF00000000CCCC.png, Context srgb, ImageData display-p3
  assert_true: Actual pixel value 255,57,40,204 is approximately equal to 234,51,35,204. expected true got false
[FAIL] Adobe-RGB-FFFF00000000CCCC.png, Context srgb, ImageData display-p3
  assert_true: Actual pixel value 255,61,43,204 is approximately equal to 234,51,35,204. expected true got false
[FAIL] Adobe-RGB-FFFF00000000CCCC.png, Context display-p3, ImageData srgb
  assert_true: Actual pixel value 255,2,0,204 is approximately equal to 255,19,11,204. expected true got false
Harness: the test ran to completion.
